backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
Pillow==10.2.0
django-environ==0.11.2
djangorestframework>=3.14
django-cors-headers>=4.0
django-filter==25.1
djangorestframework-simplejwt==5.2.2